Update: Add different files for all Keychain classes
